Expose compileSelector while retaining internal caching
